# Slightly speed up the `resize_inner` function + documentation of other functions.