@Zoxc 2023-08-18
I tested this PR out with rustc and it does seem to be a performance improvement there:
<table><tr><td rowspan="2">Benchmark</td><td colspan="1"><b>Before</b></th><td colspan="2"><b>After</b></th></tr><tr><td align="right">Time</td><td align="right">Time</td><td align="right">%</th></tr><tr><td>🟣 <b>clap</b>:check</td><td align="right">1.6201s</td><td align="right">1.6077s</td><td align="right"> -0.77%</td></tr><tr><td>🟣 <b>hyper</b>:check</td><td align="right">0.2400s</td><td align="right">0.2389s</td><td align="right"> -0.47%</td></tr><tr><td>🟣 <b>regex</b>:check</td><td align="right">0.9139s</td><td align="right">0.9061s</td><td align="right"> -0.85%</td></tr><tr><td>🟣 <b>syn</b>:check</td><td align="right">1.5023s</td><td align="right">1.4907s</td><td align="right"> -0.77%</td></tr><tr><td>🟣 <b>syntex_syntax</b>:check</td><td align="right">5.6779s</td><td align="right">5.6401s</td><td align="right"> -0.67%</td></tr><tr><td>Total</td><td align="right">9.9542s</td><td align="right">9.8835s</td><td align="right"> -0.71%</td></tr><tr><td>Summary</td><td align="right">1.0000s</td><td align="right">0.9930s</td><td align="right"> -0.70%</td></tr></table>