@JustForFun88 2023-08-20
@Amanieu I implemented `RawIter` on top of `FullBucketsIndices`, could you please review the PR?
Of course, we can go further and try to built `RawIterRange` on top of `FullBucketsIndices` too 😄 (although I haven't thought about it yet).

@Zoxc Could you please test this PR with rustc again?