@Zoxc 2023-08-20
Testing just the `RawIter` change:
<table><tr><td rowspan="2">Benchmark</td><td colspan="1"><b>Before</b></th><td colspan="2"><b>After</b></th></tr><tr><td align="right">Time</td><td align="right">Time</td><td align="right">%</th></tr><tr><td>🟣 <b>clap</b>:check</td><td align="right">1.6238s</td><td align="right">1.6322s</td><td align="right"> 0.51%</td></tr><tr><td>🟣 <b>hyper</b>:check</td><td align="right">0.2345s</td><td align="right">0.2354s</td><td align="right"> 0.37%</td></tr><tr><td>🟣 <b>regex</b>:check</td><td align="right">0.9129s</td><td align="right">0.9183s</td><td align="right"> 0.59%</td></tr><tr><td>🟣 <b>syn</b>:check</td><td align="right">1.4699s</td><td align="right">1.4747s</td><td align="right"> 0.32%</td></tr><tr><td>🟣 <b>syntex_syntax</b>:check</td><td align="right">5.6903s</td><td align="right">5.7217s</td><td align="right"> 0.55%</td></tr><tr><td>Total</td><td align="right">9.9315s</td><td align="right">9.9821s</td><td align="right"> 0.51%</td></tr><tr><td>Summary</td><td align="right">1.0000s</td><td align="right">1.0047s</td><td align="right"> 0.47%</td></tr></table>