@JustForFun88 2023-08-20
> Testing just the `RawIter` change:

I think I found the cause of the regression. Can you please repeat?