@Zoxc 2023-08-20
Still a regression:
<table><tr><td rowspan="2">Benchmark</td><td colspan="1"><b>Before</b></th><td colspan="2"><b>After</b></th></tr><tr><td align="right">Time</td><td align="right">Time</td><td align="right">%</th></tr><tr><td>🟣 <b>clap</b>:check</td><td align="right">1.6099s</td><td align="right">1.6159s</td><td align="right"> 0.37%</td></tr><tr><td>🟣 <b>hyper</b>:check</td><td align="right">0.2346s</td><td align="right">0.2356s</td><td align="right"> 0.40%</td></tr><tr><td>🟣 <b>regex</b>:check</td><td align="right">0.8979s</td><td align="right">0.9019s</td><td align="right"> 0.45%</td></tr><tr><td>🟣 <b>syn</b>:check</td><td align="right">1.4505s</td><td align="right">1.4569s</td><td align="right"> 0.44%</td></tr><tr><td>🟣 <b>syntex_syntax</b>:check</td><td align="right">5.6328s</td><td align="right">5.6628s</td><td align="right"> 0.53%</td></tr><tr><td>Total</td><td align="right">9.8257s</td><td align="right">9.8730s</td><td align="right"> 0.48%</td></tr><tr><td>Summary</td><td align="right">1.0000s</td><td align="right">1.0044s</td><td align="right"> 0.44%</td></tr></table>