@JustForFun88 2023-08-20
@Amanieu I give up 😄. As @Zoxc predicted, and then confirmed by benchmarking, the implementation of `RawIter` through `FullBucketsIndices` leads to regression. I don't think this can be changed, because in fact one operation `bucket.next_n(index)` is replaced by:
```rust
let data_index = group_first_index + index;
bucket.next_n(data_index)
```
That is, each time there is an additional `add` operation, which leads to a small but measurable regression.

To reduce the code itself, we can, of course, perform some sort of dispatching through traits and generics, like the example below, but this will likely slow down compilation and, in essence, will be equivalent to two different structures:
```rust
trait GroupBase {
    fn next_n(&self, offset: usize) -> Self;
}

impl GroupBase for usize {
    fn next_n(&self, offset: usize) -> Self {
        self + offset
    }
}

impl<T> GroupBase for Bucket<T> {
    fn next_n(&self, offset: usize) -> Self {
        unsafe { self.next_n(offset) }
    }
}

pub(crate) struct FullBucketsIndices<B: GroupBase> {
    current_group: BitMaskIter,
    group_base: B,
    // Pointer to the current group of control bytes,
    ctrl: NonNull<u8>,
    // Buckets number in given subset of a table.
    buckets: usize,
    // Number of elements in given subset of a table.
    items: usize,
}
```