@Amanieu 2023-08-22
That's fine, let's just switch back to the previous implementation of `RawIter` then.