@JustForFun88 2023-08-26
> That's fine, let's just switch back to the previous implementation of `RawIter` then.

Returned the previous implementation. However, the tests are failing, it looks like an upstrem problem [rust-lang/rust#115239](https://github.com/rust-lang/rust/issues/115239)